+++ dbc_influxdb/filetypereader.py
@@ -1,3 +1,4 @@
+import _csv
 import fnmatch
 import os
 from logging import Logger
@@ -35,6 +36,7 @@ class FileTypeReader():
         self.logger = logger
 
         self.data_df = pd.DataFrame()
+        self.df_list = pd.DataFrame()
 
         # Settings for .read_csv
 
@@ -72,9 +74,15 @@ class FileTypeReader():
         self.data_encoding = filetypeconf['data_encoding']
 
         # Special file format
-        special_formats = ['-ICOSSEQ-']
+        self.special_format = None
+        special_formats = ['-ICOSSEQ-', '-ALTERNATING-']
         for sf in special_formats:
-            self.special_format = sf if sf in self.filetype else False
+            if sf in self.filetype:
+                self.special_format = sf
+                break
+            else:
+                continue
+            # self.special_format = sf if sf in self.filetype else False
 
         self.file_info = dict(
             filepath=self.filepath,
@@ -84,21 +92,24 @@ class FileTypeReader():
 
         self.goodrows_col = None if not filetypeconf['data_keep_good_rows'] \
             else filetypeconf['data_keep_good_rows'][0]  # Col used to identify rows w/ good data
-        self.goodrows_id = None if not filetypeconf['data_keep_good_rows'] \
-            else filetypeconf['data_keep_good_rows'][1]  # ID used to identify rows w/ good data
+        self.goodrows_ids = None if not filetypeconf['data_keep_good_rows'] \
+            else filetypeconf['data_keep_good_rows'][1:]  # ID(s) used to identify rows w/ good data
 
         self.badrows_col = None if not filetypeconf['data_remove_bad_rows'] \
             else filetypeconf['data_remove_bad_rows'][0]  # Col used to identify rows w/ bad data
-        self.badrows_id = None if not filetypeconf['data_remove_bad_rows'] \
-            else filetypeconf['data_remove_bad_rows'][1]  # ID used to identify rows w/ bad data
+        self.badrows_ids = None if not filetypeconf['data_remove_bad_rows'] \
+            else filetypeconf['data_remove_bad_rows'][1:]  # ID(s) used to identify rows w/ bad data
 
         # # Detect if file has special format that needs formatting
         # self.is_special_format = True if any(f in self.config_filetype for f in self.special_formats) else False
 
+        # Config is also needed later
+        self.filetypeconf = filetypeconf
+
         self.run()
 
     def get_data(self):
-        return self.data_df, self.file_info
+        return self.df_list, self.file_info
 
     def run(self):
         self.data_df = self._readfile()
@@ -118,29 +129,43 @@ class FileTypeReader():
         self._convert_to_float_or_string()
 
         # Convert special data structures
+        # After this step, the dataframe(s) is/are stored in a list.
+        # Reason is that the special format -ALTERNATING- returns
+        # two different dataframes from one file, which are returned
+        # in a list. Other special formats are therefore also returned
+        # in a list, to be consistent.
         if self.special_format:
-            self._special_data_formats()
+            self.df_list = self._special_data_formats()
+
+        # ** After this step, self.df_list instead of self.data_df is used for further processing **
+
+        # Convert dataframe to list of dataframes if not already done
+        if not isinstance(self.df_list, list):
+            self.df_list = [self.data_df.copy()]
+        self.data_df = None  # Additional safeguard to really use df_list from now on
 
         # Filter data
-        if self.goodrows_id:
+        if self.goodrows_ids:
             self._keep_good_data_rows()
-        if self.badrows_id:
+
+        if self.badrows_ids:
             self._remove_bad_data_rows()
 
         # todo check if this works Numeric data only
         self._to_numeric()
 
         # Sanitize data
-        # Replace inf and -inf with NAN v0.7.0
+        # Replace inf and -inf with NAN
         # Sometimes inf or -inf can appear, they are interpreted
         # as some sort of number (i.e., the column dtype does not
         # become 'object' and they are not a string) but cannot be
         # handled.
-        self.data_df.replace([np.inf, -np.inf], np.nan, inplace=True)
+        for ix, df in enumerate(self.df_list):
+            self.df_list[ix].replace([np.inf, -np.inf], np.nan, inplace=True)
 
         # Timestamp
         if self.build_timestamp:
-            self.data_df = self._build_timestamp()
+            self._build_timestamp()
         self._sort_timestamp()
         self._remove_index_duplicates()
 
@@ -194,34 +219,88 @@ class FileTypeReader():
 
     def _keep_good_data_rows(self):
         """Keep good data rows only, needed for irregular formats"""
-        filter_goodrows = self.data_df.iloc[:, self.goodrows_col] == self.goodrows_id
-        self.data_df = self.data_df[filter_goodrows]
+        for ix, goodrows_id in enumerate(self.goodrows_ids):
+            filter_goodrows = self.df_list[ix].iloc[:, self.goodrows_col] == goodrows_id
+            self.df_list[ix] = self.df_list[ix][filter_goodrows].copy()
 
     def _remove_bad_data_rows(self):
         """Remove bad data rows, needed for irregular formats"""
-        filter_badrows = self.data_df.iloc[:, self.badrows_col] != self.badrows_id
-        self.data_df = self.data_df[filter_badrows]
+        for ix, badrows_id in enumerate(self.badrows_ids):
+            filter_badrows = self.df_list[ix].iloc[:, self.badrows_col] != badrows_id
+            self.df_list[ix] = self.df_list[ix][filter_badrows].copy()
 
-    def _special_data_formats(self):
+    def _special_data_formats(self) -> list:
+        """
+        Convert data with special format to conventional format
 
+        Conversion of the special format -ALTERNATING- results in
+        two different dataframes that are returned in a list. To
+        account for this special case, all other special format
+        conversions also need to be returned as lists after this
+         step for the sake of consistency.
+        """
         if self.special_format == '-ICOSSEQ-':
-            self._special_format_icosseq()
+            df = self._special_format_icosseq()  # Returns single dataframe
+        elif self.special_format == '-ALTERNATING-':
+            df = self._special_format_alternating()  # Returns two dataframes in list
+        df_list = [df] if not isinstance(df, list) else df
+        return df_list
 
-    def _special_format_icosseq(self):
-        """Convert structure of sequential ICOSSEQ files
+    def _special_format_alternating(self) -> list:
+        """Special format -ALTERNATING-
+
+        Store data with differrent IDs in different dataframes
+
+        This special format stores data from two different sources in
+        one file. Each row contains an ID that indicates from which
+        data source the respective data row originated. Typically,
+        the ID changes with each row. For example, first row has
+        ID 225, second row ID 1, third row ID 225, fourth row ID 1, etc.
+
+        However, first all file data with all different IDs is stored
+        in one dataframe. The column names in this one dataframe are
+        wrong at this point, they are fixed here.
+
+        Therefore, this dataframe is split into two dataframes here.
+        Each dataframe then contains data from one single data source
+        and all rows have the same ID. Since the two dfs have a different
+        number of vars stored in them, the column names are then also fixed.
+
+        """
+        # Get data for first ID
+        filter_data_rows = self.data_df.iloc[:, self.goodrows_col] == self.goodrows_ids[0]
+        data_df = self.data_df[filter_data_rows].copy()
+        raw_varnames = list(self.filetypeconf['data_vars'].keys())  # Varnames for this ID
+        n_raw_varnames = len(raw_varnames)  # Number of vars for this ID
+        data_df = data_df.iloc[:, 0:n_raw_varnames].copy()  # Keep number of cols for this ID
+        data_df.columns = raw_varnames  # Assign correct varnames
+
+        # Get data for second ID
+        filter_data2_rows = self.data_df.iloc[:, self.goodrows_col] == self.goodrows_ids[1]
+        data2_df = self.data_df[filter_data2_rows].copy()
+        raw_varnames = list(self.filetypeconf['data_vars2'].keys())
+        n_raw_varnames = len(raw_varnames)
+        data2_df = data2_df.iloc[:, 0:n_raw_varnames].copy()
+        data2_df.columns = raw_varnames
+
+        return [data_df, data2_df]
+
+    def _special_format_icosseq(self) -> pd.DataFrame:
+        """Convert structure of sequential -ICOSSEQ- files
 
         This file format stores different heights for the same var in different
         rows, which is a difference to regular formats. Here, the data format
         is converted in a way so that each variable for each height is in a
         separate column.
 
-        In case of ICOSSEQ files, the colname giving the height of the measurements
+        In case of -ICOSSEQ- files, the colname giving the height of the measurements
         is either LOCATION (newer files) or INLET (older files).
 
         This conversion makes sure that different heights are stored in different
         columns instead of different rows.
 
         """
+        df = self.data_df.copy()
 
         # Detect subformat: profile or chambers
         subformat = None
@@ -231,12 +310,12 @@ class FileTypeReader():
             subformat = 'CMB'
 
         # Detect name of col where the different heights are stored
-        locs_col = 'LOCATION' if any('LOCATION' in col for col in self.data_df.columns) else False
+        locs_col = 'LOCATION' if any('LOCATION' in col for col in df.columns) else False
         if not locs_col:
-            locs_col = 'INLET' if any('INLET' in col for col in self.data_df.columns) else False
+            locs_col = 'INLET' if any('INLET' in col for col in df.columns) else False
 
         # Detect unique locations identifiers, e.g. T1_35
-        locations = self.data_df[locs_col].dropna()
+        locations = df[locs_col].dropna()
         locations = locations.unique()
         locations = list(locations)
         locations.sort()
@@ -245,14 +324,14 @@ class FileTypeReader():
         # Loop over data subsets of unique locations and collect all in df
         locations_df = pd.DataFrame()
         for loc in locations:
-            _loc_df = self.data_df.loc[self.data_df[locs_col] == loc, :]
+            _loc_df = df.loc[df[locs_col] == loc, :]
 
             # If chambers, add vertical position index 0 (measured at zero height/depth)
             if subformat == 'CMB':
                 loc = f"{loc}_0" if subformat == 'CMB' else loc
 
             renamedcols = []
-            for col in self.data_df.columns:
+            for col in df.columns:
                 # _newname_suffix = '' if _newname_suffix in col else 'CMB' if subformat ==
 
                 # Add subformat info to newname,
@@ -279,40 +358,46 @@ class FileTypeReader():
         locations_df = locations_df[subsetcols]
 
         # Set the collected and converted data as main data
-        self.data_df = locations_df
+        df = locations_df
+        return df
 
     def _remove_index_duplicates(self):
-        self.data_df = self.data_df[~self.data_df.index.duplicated(keep='last')]
+        for ix, df in enumerate(self.df_list):
+            self.df_list[ix] = self.df_list[ix][~self.df_list[ix].index.duplicated(keep='last')]
 
     def _sort_timestamp(self):
-        self.data_df.sort_index(inplace=True)
+        for ix, df in enumerate(self.df_list):
+            self.df_list[ix].sort_index(inplace=True)
 
     def _remove_unnamed_cols(self):
         """Remove columns that do not have a column name"""
-        newcols = []
-        for col in self.data_df.columns:
-            if any('Unnamed' in value for value in col):
-                pass
-            else:
-                newcols.append(col)
-        self.data_df = self.data_df[newcols]
-        self.data_df.columns = pd.MultiIndex.from_tuples(newcols)
+        for ix, df in enumerate(self.df_list):
+            newcols = []
+            for col in self.df_list[ix].columns:
+                if any('Unnamed' in value for value in col):
+                    pass
+                else:
+                    newcols.append(col)
+            self.df_list[ix] = self.df_list[ix][newcols]
+            self.df_list[ix].columns = pd.MultiIndex.from_tuples(newcols)
 
     def _rename_unnamed_units(self):
         """Units not given in files with two-row header yields "Unnamed ..." units"""
-        newcols = []
-        for col in self.data_df.columns:
-            if any('Unnamed' in value for value in col):
-                col = (col[0], '-not-given-')
-            newcols.append(col)
-        self.data_df.columns = pd.MultiIndex.from_tuples(newcols)
+        for ix, df in enumerate(self.df_list):
+            newcols = []
+            for col in self.df_list[ix].columns:
+                if any('Unnamed' in value for value in col):
+                    col = (col[0], '-not-given-')
+                newcols.append(col)
+            self.df_list[ix].columns = pd.MultiIndex.from_tuples(newcols)
 
     def _to_numeric(self):
         """Make sure all data are numeric"""
-        try:
-            self.data_df = self.data_df.astype(float)  # Crashed if not possible
-        except:
-            self.data_df = self.data_df.apply(pd.to_numeric, errors='coerce')  # Does not crash
+        for ix, df in enumerate(self.df_list):
+            try:
+                self.df_list[ix] = self.df_list[ix].astype(float)  # Crashes if not possible
+            except:
+                self.df_list[ix] = self.df_list[ix].apply(pd.to_numeric, errors='coerce')  # Does not crash
 
     def _get_date_parser(self, parser):
         return lambda c: pd.to_datetime(c, format=parser, errors='coerce') if parser else False
@@ -320,13 +405,13 @@ class FileTypeReader():
 
     def _add_units_row(self):
         """Units not given in files with single-row header"""
-        if not isinstance(self.data_df.columns, pd.MultiIndex):
-            # if header and len(header) == 1:
-            newcols = []
-            for col in self.data_df.columns:
-                newcol = (col, '-not-given-')
-                newcols.append(newcol)
-            self.data_df.columns = pd.MultiIndex.from_tuples(newcols)
+        for ix, df in enumerate(self.df_list):
+            if not isinstance(self.df_list[ix].columns, pd.MultiIndex):
+                newcols = []
+                for col in self.df_list[ix].columns:
+                    newcol = (col, '-not-given-')
+                    newcols.append(newcol)
+                self.df_list[ix].columns = pd.MultiIndex.from_tuples(newcols)
 
     def _readfile(self):
         """Read data file"""
@@ -341,6 +426,7 @@ class FileTypeReader():
                     parse_dates=self.parse_dates,
                     date_parser=self.date_parser,
                     index_col=self.index_col,
+                    # engine='c',
                     engine='python',
                     # nrows=5,
                     nrows=self.nrows,
@@ -370,84 +456,88 @@ class FileTypeReader():
             # and then the erroneous row is skipped (1).
             args['skiprows'] = [0, 1]
             df = pd.read_csv(**args)
+        except _csv.Error:
+            # The _csv.Error occurs e.g. in case there are NUL bytes in
+            # the data file. The python engine cannot handle these bytes,
+            # but the c engine can.
+            args['engine'] = 'c'
+            df = pd.read_csv(**args)
 
         return df
 
     def _build_timestamp(self):
         """
         Build full datetime timestamp by combining several cols
-
-        :param df:
-        :param method:
-        :return:
-        pd.DataFrame
         """
 
-        df = self.data_df.copy()
+        df_list = self.df_list
 
-        # Build from columns by index, column names not available
-        if self.build_timestamp == 'YEAR0+MONTH1+DAY2+HOUR3+MINUTE4':
-            # Remove rows where date info is missing
-            _not_possible = df['YEAR'].isnull()
-            df = df[~_not_possible]
-            _not_possible = df['MONTH'].isnull()
-            df = df[~_not_possible]
-            _not_possible = df['DAY'].isnull()
-            df = df[~_not_possible]
-            _not_possible = df['HOUR'].isnull()
-            df = df[~_not_possible]
-            _not_possible = df['MINUTE'].isnull()
-            df = df[~_not_possible]
+        for ix, df in enumerate(df_list):
 
-            # pandas recognizes columns with these names as time columns
-            df['TIMESTAMP'] = pd.to_datetime(df[['YEAR', 'MONTH', 'DAY', 'HOUR', 'MINUTE']])
+            # Build from columns by index, column names not available
+            if self.build_timestamp == 'YEAR0+MONTH1+DAY2+HOUR3+MINUTE4':
+                # Remove rows where date info is missing
+                _not_possible = df['YEAR'].isnull()
+                df = df[~_not_possible]
+                _not_possible = df['MONTH'].isnull()
+                df = df[~_not_possible]
+                _not_possible = df['DAY'].isnull()
+                df = df[~_not_possible]
+                _not_possible = df['HOUR'].isnull()
+                df = df[~_not_possible]
+                _not_possible = df['MINUTE'].isnull()
+                df = df[~_not_possible]
 
-            # Remove original columns
-            dropcols = ['YEAR', 'MONTH', 'DAY', 'HOUR', 'MINUTE']
-            df.drop(dropcols, axis=1, inplace=True)
+                # pandas recognizes columns with these names as time columns
+                df['TIMESTAMP'] = pd.to_datetime(df[['YEAR', 'MONTH', 'DAY', 'HOUR', 'MINUTE']])
 
-            # Remove rows where timestamp-building did not work
-            locs_emptydate = df['TIMESTAMP'].isnull()
-            df = df.loc[~locs_emptydate, :]
+                # Remove original columns
+                dropcols = ['YEAR', 'MONTH', 'DAY', 'HOUR', 'MINUTE']
+                df.drop(dropcols, axis=1, inplace=True)
 
-            # Set as index
-            df.set_index('TIMESTAMP', inplace=True)
+                # Remove rows where timestamp-building did not work
+                locs_emptydate = df['TIMESTAMP'].isnull()
+                df = df.loc[~locs_emptydate, :]
 
-            return df
+                # Set as index
+                df.set_index('TIMESTAMP', inplace=True)
 
-        # Build from columns by name, column names available
-        if self.build_timestamp == 'YEAR+DOY+TIME':
-            # Remove rows where date info is missing
-            _not_possible = df['YEAR'].isnull()
-            df = df[~_not_possible]
-            _not_possible = df['DOY'].isnull()
-            df = df[~_not_possible]
-            _not_possible = df['DOY'] == 0
-            df = df[~_not_possible]
-            _not_possible = df['TIME'].isnull()
-            df = df[~_not_possible]
+                self.df_list[ix] = df.copy()
 
-            df['_basedate'] = pd.to_datetime(df['YEAR'], format='%Y', errors='coerce')
-            df['_doy_timedelta'] = pd.to_timedelta(df['DOY'], unit='D') - pd.Timedelta(days=1)
-            df['_time_str'] = df['TIME'].astype(int).astype(str).str.zfill(4)
-            df['_time'] = pd.to_datetime(df['_time_str'], format='%H%M', errors='coerce')
-            df['_hours'] = df['_time'].dt.hour
-            df['_hours'] = pd.to_timedelta(df['_hours'], unit='hours')
-            df['_minutes'] = df['_time'].dt.minute
-            df['_minutes'] = pd.to_timedelta(df['_minutes'], unit='minutes')
+            # Build from columns by name, column names available
+            if self.build_timestamp == 'YEAR+DOY+TIME':
+                # Remove rows where date info is missing
+                _not_possible = df['YEAR'].isnull()
+                df = df[~_not_possible]
+                _not_possible = df['DOY'].isnull()
+                df = df[~_not_possible]
+                _not_possible = df['DOY'] == 0
+                df = df[~_not_possible]
+                _not_possible = df['TIME'].isnull()
+                df = df[~_not_possible]
 
-            df['TIMESTAMP'] = df['_basedate'] \
-                              + df['_doy_timedelta'] \
-                              + df['_hours'] \
-                              + df['_minutes']
+                df['_basedate'] = pd.to_datetime(df['YEAR'], format='%Y', errors='coerce')
+                df['_doy_timedelta'] = pd.to_timedelta(df['DOY'], unit='D') - pd.Timedelta(days=1)
+                df['_time_str'] = df['TIME'].astype(int).astype(str).str.zfill(4)
+                df['_time'] = pd.to_datetime(df['_time_str'], format='%H%M', errors='coerce')
+                df['_hours'] = df['_time'].dt.hour
+                df['_hours'] = pd.to_timedelta(df['_hours'], unit='hours')
+                df['_minutes'] = df['_time'].dt.minute
+                df['_minutes'] = pd.to_timedelta(df['_minutes'], unit='minutes')
 
-            dropcols = ['_basedate', '_doy_timedelta', '_hours', '_minutes', '_time', '_time_str',
-                        'YEAR', 'DOY', 'TIME']
-            df.drop(dropcols, axis=1, inplace=True)
-            locs_emptydate = df['TIMESTAMP'].isnull()
-            df = df.loc[~locs_emptydate, :]
-            df.set_index('TIMESTAMP', inplace=True)
-            return df
+                df['TIMESTAMP'] = df['_basedate'] \
+                                  + df['_doy_timedelta'] \
+                                  + df['_hours'] \
+                                  + df['_minutes']
+
+                dropcols = ['_basedate', '_doy_timedelta', '_hours', '_minutes', '_time', '_time_str',
+                            'YEAR', 'DOY', 'TIME']
+                df.drop(dropcols, axis=1, inplace=True)
+                locs_emptydate = df['TIMESTAMP'].isnull()
+                df = df.loc[~locs_emptydate, :]
+                df.set_index('TIMESTAMP', inplace=True)
+
+                self.df_list[ix] = df.copy()
 
 
 def get_conf_filetypes(dir: Path, ext: str = 'yaml') -> dict: