+++ dbc_influxdb/manual_run.py
@@ -1,4 +1,5 @@
 from pathlib import Path
+
 from dbc_influxdb.main import dbcInflux
 
 
@@ -7,8 +8,7 @@ def upload_specific_file():
     Upload specific file to database
     """
 
-    # to_bucket = 'test'
-    dir_base = r'F:\Dropbox\luhk_work\40 - DATA\FLUXNET-WW2020_RELEASE-2022-1'
+    # dir_base = r'F:\Dropbox\luhk_work\40 - DATA\FLUXNET-WW2020_RELEASE-2022-1'
 
     # CH-AWS
     # to_bucket = 'ch-aws_processing'
@@ -35,16 +35,26 @@ def upload_specific_file():
     # dir_site = r'FLX_CH-Lae_FLUXNET2015_FULLSET_2004-2020_beta-3'
     # file_site = r'FLX_CH-Lae_FLUXNET2015_FULLSET_HH_2004-2020_beta-3.csv'
 
-    # CH-OE2
-    to_bucket = 'ch-oe2_processing'
-    dir_site = r'FLX_CH-Oe2_FLUXNET2015_FULLSET_2004-2020_beta-3'
-    file_site = r'FLX_CH-Oe2_FLUXNET2015_FULLSET_HH_2004-2020_beta-3.csv'
+    # # CH-OE2
+    # to_bucket = 'ch-oe2_processing'
+    # dir_site = r'FLX_CH-Oe2_FLUXNET2015_FULLSET_2004-2020_beta-3'
+    # file_site = r'FLX_CH-Oe2_FLUXNET2015_FULLSET_HH_2004-2020_beta-3.csv'
+
+    to_bucket = 'test'
+    dir_base = r'F:\Downloads'
+    dir_site = r'_temp'
+    file_site = r'CH-DAV_iDL_T1_35_1_TBL1_2022_10_25_0000.dat.csv'
 
     # Prepare upload settings
     filepath = str(Path(dir_base) / Path(dir_site) / file_site)
-    data_version = 'FLUXNET-WW2020_RELEASE-2022-1'  # Important tag
+
+    # data_version = 'FLUXNET-WW2020_RELEASE-2022-1'  # Important tag
+    data_version = 'raw'  # Important tag
+
     dirconf = r'F:\Dropbox\luhk_work\20 - CODING\22 - POET\configs'  # Configurations
-    filetype = 'PROC-FLUXNET-FULLSET-HH-CSV-30MIN'
+
+    # filetype = 'PROC-FLUXNET-FULLSET-HH-CSV-30MIN'
+    filetype = 'DAV10-RAW-TBL1-201802281101-TOA5-DAT-10S'
 
     # Instantiate class
     dbc = dbcInflux(dirconf=dirconf)
@@ -73,20 +83,15 @@ def download():
     """
 
     # Settings
-    BUCKET = 'ch-dav_processing'
-    MEASUREMENTS=['TA', 'RH', 'SW', 'PPFD', 'LW', 'PA']
+    BUCKET = 'ch-fru_raw'
+    MEASUREMENTS = ['SW']
     FIELDS = [
-        # 'TA_NABEL_T1_35_1',
-        # 'RH_NABEL_T1_35_1',
-        # 'SW_IN_NABEL_T1_35_1',
-        # 'LW_IN_T1_35_2',
-        'PPFD_IN_T1_35_2',
-        'PA_H1_0_1'
+        'SW_IN_T1_1_1',
     ]
-    START = '2021-01-01 00:01:00'
-    STOP = '2023-01-01 00:01:00'
+    START = '2022-03-01 00:00:00'
+    STOP = '2022-05-01 00:00:10'
     TIMEZONE_OFFSET_TO_UTC_HOURS = 1  # We need returned timestamps in CET (winter time), which is UTC + 1 hour
-    DATA_VERSION='meteoscreening'
+    DATA_VERSION = 'raw'
     DIRCONF = r'L:\Dropbox\luhk_work\20 - CODING\22 - POET\configs'
 
     # Instantiate class
@@ -104,7 +109,7 @@ def download():
             data_version=DATA_VERSION
         )
 
-    # data_simple.to_csv("M:\Downloads\_temp\del.csv")
+    data_simple.to_csv("M:\Downloads\_temp\del.csv")
     print(data_simple)